GroupServiceImpl -> GroupService. UISelect.createGroupSelect fixed (used e.g. by UserPagesRest)
