+++ projectforge-business/src/test/java/org/projectforge/business/group/service/GroupsProviderTest.java
@@ -46,7 +46,7 @@ public class GroupsProviderTest
   @Test
   public void convertGroupIds()
   {
-    final GroupServiceImpl groupService = new GroupServiceImpl();
+    final GroupService groupService = new GroupService();
     groupService.setUserGroupCache(Mockito.mock(UserGroupCache.class));
     groupService.setGroupDao(Mockito.mock(GroupDao.class));
     Mockito.when(groupService.getGroup(1L)).thenReturn(cg("1", 1));
